requirements.txt
# Core FastAPI framework
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# LangChain RAG system components
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# External integrations (optional for demo)
sendgrid>=6.0.0
google-api-python-client>=2.100.0
google-auth>=2.0.0
python-dotenv>=1.0.0

# Additional utilities
python-multipart>=0.0.6